add: more unit test + test cleanup
