+++ package.json
@@ -30,6 +30,7 @@
   "license": "MIT",
   "devDependencies": {
     "csstype": "^3.1.2",
+    "jsdom": "^22.1.0",
     "prettier": "2.8",
     "tsup": "^7.1.0",
     "typescript": "^5.1.6",
